# spec/openapi/plugin-platform-v7.yaml
openapi: 3.1.0
info:
  title: ReDevPlugin Plugin Platform API
  version: plugin-platform-v7
  description: Host-agnostic plugin management, bridge, operation, data, and sandbox routes. Release-ref install and update routes resolve host-controlled package artifacts server-side and do not accept requested trust_state values; local-import package install/update routes are explicit developer/import flows and also do not accept caller-supplied trust_state values.
...
paths:
  /_redevplugin/api/plugins/{plugin_instance_id}/local-import:
    post:
      operationId: importLocalPackage
      x-redevplugin-route-effect: mutation
      parameters:
        - $ref: "#/components/parameters/PluginInstanceID"
      requestBody:
        required: true
        content:
          application/vnd.redevplugin.package+zip:
            schema: { type: string, format: binary }
      responses:
        "200": { $ref: "#/components/responses/PluginRecordResponse" }
        default: { $ref: "#/components/responses/MutationPlatformErrorResponse" }
    put:
      operationId: updateLocalPackage
      x-redevplugin-route-effect: mutation
      parameters:
        - $ref: "#/components/parameters/PluginInstanceID"
        - name: X-ReDevPlugin-Expected-Management-Revision
          in: header
          required: true
          description: Canonical decimal management revision. Duplicate, combined, empty, or whitespace-padded values are rejected.
          schema: { type: integer, minimum: 1, maximum: 9007199254740991 }
      requestBody:
        required: true
        content:
          application/vnd.redevplugin.package+zip:
            schema: { type: string, format: binary }
      responses:
        "200": { $ref: "#/components/responses/PluginRecordResponse" }
        default: { $ref: "#/components/responses/MutationPlatformErrorResponse" }
...
  /_redevplugin/api/plugins/enable:
    post:
      operationId: enablePlugin
      x-redevplugin-route-effect: mutation
      requestBody:
        $ref: "#/components/requestBodies/EnableRequest"
      responses:
        "200": { $ref: "#/components/responses/PluginRecordResponse" }
        default: { $ref: "#/components/responses/MutationPlatformErrorResponse" }
...
  /_redevplugin/api/plugins/catalog/query:
    post:
      operationId: listPlugins
      x-redevplugin-route-effect: query
      requestBody:
        $ref: "#/components/requestBodies/EmptyQueryRequest"
      responses:
        "200": { $ref: "#/components/responses/PluginCatalogResponse" }
        default: { $ref: "#/components/responses/PlatformErrorResponse" }
...
components:
  parameters:
...
    PluginInstanceID:
      name: plugin_instance_id
      in: path
      required: true
      schema: { type: string }
...
  responses:
    PlatformErrorResponse:
      description: Closed non-success response for read-only platform operations.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PlatformErrorResponse"
    MutationPlatformErrorResponse:
      description: Closed non-success response for operations that may change durable platform state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MutationPlatformErrorResponse"
...
    PluginRecordResponse:
      description: Plugin lifecycle response with typed registry trust assessment.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PluginRecordSuccessResponse"
    PluginCatalogResponse:
      description: Installed plugin catalog response with typed registry records.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PluginCatalogSuccessResponse"
...
  requestBodies:
...
    EnableRequest:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/EnableRequest"
...
    EmptyQueryRequest:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/EmptyQueryRequest"
...
  schemas:
    PlatformErrorResponse:
      type: object
      required: [ok, error]
      additionalProperties: false
      properties:
        ok: { const: false }
        error:
          $ref: "#/components/schemas/PlatformError"
    MutationPlatformErrorResponse:
      type: object
      required: [ok, error]
      additionalProperties: false
      properties:
        ok: { const: false }
        error:
          $ref: "#/components/schemas/MutationPlatformError"
...
    PluginRecordSuccessResponse:
      type: object
      required: [ok, data]
      additionalProperties: false
      properties:
        ok: { const: true }
        data: { $ref: "#/components/schemas/PluginRecord" }
    PluginCatalogSuccessResponse:
      type: object
      required: [ok, data]
      additionalProperties: false
      properties:
        ok: { const: true }
        data: { $ref: "#/components/schemas/PluginCatalogResult" }
...
    PlatformError:
      oneOf:
        - $ref: "#/components/schemas/GenericPlatformError"
        - $ref: "#/components/schemas/JSONLimitPlatformError"
        - $ref: "#/components/schemas/PackageValidationPlatformError"
        - $ref: "#/components/schemas/ManagementRevisionPlatformError"
        - $ref: "#/components/schemas/AuthorizationRevisionPlatformError"
        - $ref: "#/components/schemas/BindingRevisionPlatformError"
        - $ref: "#/components/schemas/ValuesRevisionPlatformError"
        - $ref: "#/components/schemas/CapabilityPlatformError"
        - $ref: "#/components/schemas/WorkerPlatformError"
    MutationPlatformError:
      oneOf:
        - $ref: "#/components/schemas/MutationGenericPlatformError"
        - $ref: "#/components/schemas/MutationJSONLimitPlatformError"
        - $ref: "#/components/schemas/MutationPackageValidationPlatformError"
        - $ref: "#/components/schemas/MutationManagementRevisionPlatformError"
        - $ref: "#/components/schemas/MutationAuthorizationRevisionPlatformError"
        - $ref: "#/components/schemas/MutationBindingRevisionPlatformError"
        - $ref: "#/components/schemas/MutationValuesRevisionPlatformError"
        - $ref: "#/components/schemas/MutationCapabilityPlatformError"
        - $ref: "#/components/schemas/MutationWorkerPlatformError"
    GenericPlatformError:
      type: object
      required: [code, message, details]
      additionalProperties: false
      properties:
        code:
          $ref: "#/components/schemas/GenericPlatformErrorCode"
        message: { $ref: "#/components/schemas/PlatformErrorMessage" }
        details:
          $ref: "#/components/schemas/EmptyErrorDetails"
...
    MutationGenericPlatformError:
      type: object
      required: [code, message, details, mutation_outcome]
      additionalProperties: false
      properties:
        code: { $ref: "#/components/schemas/GenericPlatformErrorCode" }
        message: { $ref: "#/components/schemas/PlatformErrorMessage" }
        details: { $ref: "#/components/schemas/EmptyErrorDetails" }
        mutation_outcome: { $ref: "#/components/schemas/MutationOutcome" }
    MutationJSONLimitPlatformError:
      type: object
      required: [code, message, details, mutation_outcome]
      additionalProperties: false
      properties:
        code: { const: PLUGIN_JSON_LIMIT_EXCEEDED }
        message: { $ref: "#/components/schemas/PlatformErrorMessage" }
        details: { $ref: "#/components/schemas/JSONLimitErrorDetails" }
        mutation_outcome: { $ref: "#/components/schemas/MutationOutcome" }
    MutationPackageValidationPlatformError:
      type: object
      required: [code, message, details, mutation_outcome]
      additionalProperties: false
      properties:
        code:
          type: string
          enum: [PLUGIN_MANIFEST_INVALID, PLUGIN_PACKAGE_INVALID, PLUGIN_PACKAGE_TOO_LARGE, PLUGIN_PACKAGE_PATH_FORBIDDEN]
        message: { $ref: "#/components/schemas/PlatformErrorMessage" }
        details: { $ref: "#/components/schemas/PackageValidationErrorDetails" }
        mutation_outcome: { $ref: "#/components/schemas/MutationOutcome" }
    MutationManagementRevisionPlatformError:
      type: object
      required: [code, message, details, mutation_outcome]
      additionalProperties: false
      properties:
        code: { const: PLUGIN_MANAGEMENT_REVISION_MISMATCH }
        message: { $ref: "#/components/schemas/PlatformErrorMessage" }
        details: { $ref: "#/components/schemas/ManagementRevisionMismatchDetails" }
        mutation_outcome: { $ref: "#/components/schemas/MutationOutcome" }
    MutationAuthorizationRevisionPlatformError:
      type: object
      required: [code, message, details, mutation_outcome]
      additionalProperties: false
      properties:
        code: { const: PLUGIN_AUTHORIZATION_REVISION_MISMATCH }
        message: { $ref: "#/components/schemas/PlatformErrorMessage" }
        details: { $ref: "#/components/schemas/AuthorizationRevisionMismatchDetails" }
        mutation_outcome: { $ref: "#/components/schemas/MutationOutcome" }
    MutationBindingRevisionPlatformError:
      type: object
      required: [code, message, details, mutation_outcome]
      additionalProperties: false
      properties:
        code: { const: PLUGIN_BINDING_REVISION_MISMATCH }
        message: { $ref: "#/components/schemas/PlatformErrorMessage" }
        details: { $ref: "#/components/schemas/BindingRevisionMismatchDetails" }
        mutation_outcome: { $ref: "#/components/schemas/MutationOutcome" }
    MutationValuesRevisionPlatformError:
      type: object
      required: [code, message, details, mutation_outcome]
      additionalProperties: false
      properties:
        code: { const: PLUGIN_VALUES_REVISION_MISMATCH }
        message: { $ref: "#/components/schemas/PlatformErrorMessage" }
        details: { $ref: "#/components/schemas/ValuesRevisionMismatchDetails" }
        mutation_outcome: { $ref: "#/components/schemas/MutationOutcome" }
    MutationCapabilityPlatformError:
      type: object
      required: [code, message, details, mutation_outcome]
      additionalProperties: false
      properties:
        code: { const: PLUGIN_CAPABILITY_ERROR }
        message: { $ref: "#/components/schemas/PlatformErrorMessage" }
        details: { $ref: "#/components/schemas/CapabilityBusinessErrorDetails" }
        mutation_outcome: { $ref: "#/components/schemas/MutationOutcome" }
    MutationWorkerPlatformError:
      type: object
      required: [code, message, details, mutation_outcome]
      additionalProperties: false
      properties:
        code: { const: PLUGIN_WORKER_ERROR }
        message: { $ref: "#/components/schemas/PlatformErrorMessage" }
        details: { $ref: "#/components/schemas/WorkerErrorDetails" }
        mutation_outcome: { $ref: "#/components/schemas/MutationOutcome" }
...
    PluginCatalogResult:
      type: object
      required: [plugins]
      additionalProperties: false
      properties:
        plugins:
          type: array
          items: { $ref: "#/components/schemas/PluginRecord" }
...
    StrictSemVer:
      type: string
      pattern: '^(0|[1-9][0-9]*)\.(0|[1-9][0-9]*)\.(0|[1-9][0-9]*)(?:-(?:(?:0|[1-9][0-9]*|[0-9A-Za-z-]*[A-Za-z-][0-9A-Za-z-]*)(?:\.(?:0|[1-9][0-9]*|[0-9A-Za-z-]*[A-Za-z-][0-9A-Za-z-]*))*))?(?:\+[0-9A-Za-z-]+(?:\.[0-9A-Za-z-]+)*)?$'
...
    TrustState:
      type: string
      enum: [verified, unsigned_local, untrusted, needs_review, trust_unavailable, blocked_security]
...
    TrustAssessment:
      type: object
      required: [trust_state, verified_hashes]
      additionalProperties: false
      properties:
        trust_state:
          $ref: "#/components/schemas/TrustState"
        reason_codes:
          type: array
          items: { type: string }
        verified_hashes:
          $ref: "#/components/schemas/TrustHashSet"
        verified_signature:
          $ref: "#/components/schemas/VerifiedSignature"
        trust_assessment_epoch: { type: string }
        policy_epoch: { type: string }
        revocation_epoch: { type: string }
        metadata:
          type: object
          additionalProperties: { type: string }
    PluginVersion:
      type: object
      required: [version, active_fingerprint, package_hash, manifest_hash, entries_hash, trust_state, trust_assessment, manifest, package_entries, activated_at]
      additionalProperties: false
      properties:
        version: { $ref: "#/components/schemas/StrictSemVer" }
        active_fingerprint: { type: string, minLength: 1 }
        package_hash: { type: string, minLength: 1 }
        manifest_hash: { type: string, minLength: 1 }
        entries_hash: { type: string, minLength: 1 }
        trust_state:
          $ref: "#/components/schemas/TrustState"
        trust_assessment:
          $ref: "#/components/schemas/TrustAssessment"
        source_policy_snapshot_hash: { type: string, pattern: "^(sha256:)?[0-9a-f]{64}$" }
        source_policy_snapshot:
          type: object
          additionalProperties: true
        local_import_provenance:
          $ref: "#/components/schemas/LocalImportProvenance"
        capability_contracts:
          type: array
          items: { $ref: "#/components/schemas/CapabilityContractPin" }
        manifest:
          $ref: "../plugin/manifest-v5.schema.json"
        package_entries:
          type: array
          items: { $ref: "#/components/schemas/PackageEntry" }
        runtime_requirement: { $ref: "#/components/schemas/RuntimeRequirement" }
        activated_at: { type: string, format: date-time }
        metadata:
          type: object
          additionalProperties: { type: string }
    PluginRecord:
      type: object
      description: Explicit public HTTP projection of a registry record. Persistent owner identity is excluded.
      required: [plugin_instance_id, publisher_id, plugin_id, version, active_fingerprint, package_hash, manifest_hash, entries_hash, trust_state, trust_assessment, enable_state, policy_revision, management_revision, revoke_epoch, manifest, package_entries, installed_at, updated_at]
      additionalProperties: false
      properties:
        plugin_instance_id: { type: string, minLength: 1 }
        publisher_id: { type: string }
        plugin_id: { type: string, minLength: 1 }
        version: { $ref: "#/components/schemas/StrictSemVer" }
        active_fingerprint: { type: string, minLength: 1 }
        package_hash: { type: string }
        manifest_hash: { type: string }
        entries_hash: { type: string }
        trust_state:
          $ref: "#/components/schemas/TrustState"
        trust_assessment:
          $ref: "#/components/schemas/TrustAssessment"
        source_policy_snapshot_hash: { type: string, pattern: "^(sha256:)?[0-9a-f]{64}$" }
        source_policy_snapshot:
          type: object
          additionalProperties: true
        local_import_provenance:
          $ref: "#/components/schemas/LocalImportProvenance"
        capability_contracts:
          type: array
          items: { $ref: "#/components/schemas/CapabilityContractPin" }
        enable_state:
          type: string
          enum: [disabled, enabled, disabled_by_policy, disabled_incompatible]
        disabled_reason: { type: string }
        policy_revision: { type: integer, minimum: 0, maximum: 9007199254740991 }
        management_revision: { type: integer, minimum: 0, maximum: 9007199254740991 }
        revoke_epoch: { type: integer, minimum: 0, maximum: 9007199254740991 }
        manifest:
          $ref: "../plugin/manifest-v5.schema.json"
        package_entries:
          type: array
          items: { $ref: "#/components/schemas/PackageEntry" }
        runtime_requirement: { $ref: "#/components/schemas/RuntimeRequirement" }
        version_history:
          type: array
          items:
            $ref: "#/components/schemas/PluginVersion"
        installed_at: { type: string, format: date-time }
        enabled_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        deleted_at: { type: string, format: date-time }
        metadata:
          type: object
          additionalProperties: { type: string }
    PackageEntry:
      type: object
      required: [path, size, sha256, mode]
      additionalProperties: false
      properties:
        path: { type: string, minLength: 1 }
        size: { type: integer, minimum: 0 }
        sha256: { type: string, pattern: "^[0-9a-f]{64}$" }
        mode: { type: string, minLength: 1 }
        content_type: { type: string }
    RuntimeRequirement:
      type: object
      required: [min_version]
      additionalProperties: false
      properties:
        min_version: { $ref: "#/components/schemas/StrictSemVer" }
        supported_targets:
          type: array
          uniqueItems: true
          items:
            type: string
            enum: [darwin/amd64, darwin/arm64, linux/amd64, linux/arm64]
    LocalImportProvenance:
      type: object
      required: [import_id, distribution, policy_epoch, unsigned_policy, assessed_at]
      additionalProperties: false
      properties:
        import_id: { type: string, minLength: 1 }
        distribution: { enum: [local_import] }
        policy_epoch: { type: string, minLength: 1 }
        unsigned_policy: { enum: [dev_only, review_required, block] }
        assessed_at: { type: string, format: date-time }
...
    EnableRequest:
      type: object
      required: [plugin_instance_id, expected_management_revision]
      additionalProperties: false
      properties:
        plugin_instance_id: { type: string, minLength: 1 }
        expected_management_revision: { type: integer, minimum: 1, maximum: 9007199254740991 }
...
    EmptyQueryRequest:
      type: object
      additionalProperties: false
      maxProperties: 0
...
    CapabilityContractPin:
      $ref: "../plugin/host-capability-pin-v1.schema.json"
...